implement usage of tagTranslator in remote computation environments
like on ignite, one can't re-use an established jdbc connection to a keytables database inside map, flatMap or aggregateBy lambdas, because this code is executed on remote machines which have to open their own connections to the respective keytables dbs.
this is implemented using a special implementation of a LayzEvaluatedObject (with synchronized "get"), that opens one connection to a keytables db per request (and closes is accordinly when the request is done).