use oshdb streaming api for large geojson requests
streaming of results means that features are sent out to the clients as soon as they come in from the oshdb. Which means that the requesting user gets the data earlier (downloads start "immediately"), and there is less requirements on the server to temporarily buffer the whole result set until the download is completed.

when run on ignite this does a preflight to get an estimate about if using the actual streaming api is actually needed (currently, only the "AffinityCall" mode supports it really and is a bit slower due to the required roundtrip overhead for each requested data cell. The limit is currently set somewhat arbitrarily at 10MB of osm raw data (typically corresponds to ~20MB-ish of geojson geometry data)