don't drag along non serializable objects in ProcessingData/InputProcessor objects
The HttpServletReqest objects contain a lot of internal state and references to objects that are not relevant for the use in the ohsome api.

This caused some of the ohsome api's requests to not work on ignite, which requires all used objects to be serializable, which some of the internal servelet stuff isn't.

This is fiex by refactoring the respective classes to only hold what is needed (e.g. the request URL, the request parameters and so on).