delay tag translation
when using remote tag translation with the affinitycall backend, each cell got it's own tagtranslator object and connection. This made the tagtranslator caches ineffective, causing a bottleneck and overloaded the keytables database.
doing the translation of tags on the client also has the benefit of minimizing the network traffic inside the cluster nodes